requirements.txt
absl-py==2.3.1
aiofiles==25.1.0
chex==0.1.91
einops==0.8.1
etils==1.13.0
flax==0.12.0
fsspec==2025.9.0
humanize==4.14.0
immutabledict==4.2.2
importlib_resources==6.5.2
jax==0.8.0
jaxlib==0.8.0
jaxtyping==0.3.3
markdown-it-py==4.0.0
mdurl==0.1.2
ml_dtypes==0.5.3
mock==5.2.0
msgpack==1.1.2
nest-asyncio==1.6.0
numpy==2.3.4
opt_einsum==3.4.0
optax==0.2.6
orbax-checkpoint==0.11.26
protobuf==6.33.0
psutil==7.1.2
Pygments==2.19.2
PyYAML==6.0.3
rich==14.2.0
scipy==1.16.2
simplejson==3.20.2
tensorstore==0.1.78
toolz==1.1.0
treescope==0.1.10
typeguard==4.4.4
typing_extensions==4.15.0
wadler_lindig==0.1.7
zipp==3.23.0
